+++ website/docs/components/accordion/examples/base.js
@@ -8,7 +8,7 @@ const Demo = () => (
     {[...new Array(3)].map((_, index) => (
       <Accordion.Item value={index} key={index}>
         <Accordion.Item.Toggle tag={Flex} p="8px 12px" disabled={index === 2} alignItems="center">
-          <Accordion.Item.Chevron color="stone" mr={2} />
+          <Accordion.Item.Chevron color="mist" mr={2} />
           <Text size={200} color="gray20">{`Section ${index + 1}`}</Text>
         </Accordion.Item.Toggle>
         <Accordion.Item.Collapse>