+++ website/docs/data-display/venn-chart/examples/venn.js
@@ -1,15 +1,15 @@
 import React from 'react';
-import { Plot, Venn, Tooltip } from '@semcore/d3-chart';
+import { Plot, Venn, Tooltip, colors } from '@semcore/d3-chart';
 import { Text } from '@semcore/typography';
 
 export default () => {
   return (
     <Plot height={300} width={400} data={data}>
       <Venn>
-        <Venn.Circle dataKey="G" name="Good" color="#3AB011" />
-        <Venn.Circle dataKey="F" name="Fast" color="#50AEF4" />
-        <Venn.Circle dataKey="C" name="Cheap" color="#FF8E29" />
-        <Venn.Circle dataKey="U" name="Unknown" color="#890C85" />
+        <Venn.Circle dataKey="G" name="Good" />
+        <Venn.Circle dataKey="F" name="Fast" color={colors['blue-03']} />
+        <Venn.Circle dataKey="C" name="Cheap" color={colors['orange-04']} />
+        <Venn.Circle dataKey="U" name="Unknown" color={colors['pink-03']} />
         <Venn.Intersection dataKey="G/F" name="Good & Fast" />
         <Venn.Intersection dataKey="G/C" name="Good & Cheap" />
         <Venn.Intersection dataKey="F/C" name="Fast & Cheap" />