+++ semcore/chart/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+
+## [4.3.1] - 2021-12-08
+
+### Changed
+
+- Moved chart colors vars to style
+
 ## [4.3.0] - unreleased
 
 ### Changed