+++ semcore/d3-chart/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
+## [1.5.6] - 2021-12-08
+
+### Changed
+
+- Moved chart colors vars to style
+
 ## [1.5.5] - 2021-12-06
 
 ### Fixed